+++ include/btrstd/containers/owning_array_list.h
@@ -65,6 +65,9 @@ size_t BTR_OAList_len(const btr_oalist_s *);
 bool BTR_OAList_isEmpty(const btr_oalist_s *);
 // Reverses the order of items in the borrowing array list.
 void BTR_OAList_reverse(btr_oalist_s *);
+void BTR_OAList_reserve(btr_oalist_s *, size_t itemCount);
+void BTR_OAList_reserveNew(btr_oalist_s *, size_t itemCount);
+void BTR_OAList_cropCapacity(btr_oalist_s *);
 // Deallocates the owning array list.
 void BTR_OAList_free(btr_oalist_s *);
 // Deallocates the owning array list and initializes it again with as empty.