+++ HISTORY.md
@@ -1,9 +1,15 @@
 # Update 1.2.0 (Not out yet)
 
 ### Additions
-- Added `BTR_markedForRemoval` marco in `btrstd/error.h` module.
 - Added `btrstd/string.h` module.
 - Added `btrstd/utf8.h` module.
+- Added `BTR_BAList_reserve` method in `btrstd/containers/borrowing_array_list.h`
+- Added `BTR_BAList_reserveNew` method in `btrstd/containers/borrowing_array_list.h`
+- Added `BTR_BAList_cropCapacity` method in `btrstd/containers/borrowing_array_list.h`
+- Added `BTR_OAList_reserve` method in `btrstd/containers/owning_array_list.h`
+- Added `BTR_OAList_reserveNew` method in `btrstd/containers/owning_array_list.h`
+- Added `BTR_OAList_cropCapacity` method in `btrstd/containers/owning_array_list.h`
+- Added `BTR_markedForRemoval` marco in `btrstd/error.h` module.
 
 # Update 1.1.0